tests/requirements.txt
pytest>=9.0.2
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
iniconfig>=2.1.0
pluggy>=1.6.0
colorama>=0.4.6
click>=8.3.1
Pygments>=2.19.2
isort>=5.13.2
anyio>=4.3.0
annotated-doc>=0.0.4
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
annotated-types==0.7.0
asyncpg==0.31.0
celery==5.4.0
fastapi==0.128.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pytest==9.0.2
python-dotenv==1.2.1
redis==5.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
yarl==1.22.0